// src/store/enums/OrderActionsEnum.ts
export enum OrderActionsTypes {
  GET_ALL_ORDERS = "GET_ALL_ORDERS",
  GET_ALL_ORDERS_SUCCESS = "GET_ALL_ORDERS_SUCCESS",
  GET_ALL_ORDERS_FAILURE = "GET_ALL_ORDERS_FAILURE",
  GET_ORDER = "GET_ORDER",
  GET_ORDER_SUCCESS = "GET_ORDER_SUCCESS",
  GET_ORDER_FAILURE = "GET_ORDER_FAILURE",
  GET_ORDERS_BY_USER = "GET_ORDERS_BY_USER",
  GET_ORDERS_BY_USER_SUCCESS = "GET_ORDERS_BY_USER_SUCCESS",
  GET_ORDERS_BY_USER_FAILURE = "GET_ORDERS_BY_USER_FAILURE",
  CREATE_ORDER = "CREATE_ORDER",
  CREATE_ORDER_SUCCESS = "CREATE_ORDER_SUCCESS",
  CREATE_ORDER_FAILURE = "CREATE_ORDER_FAILURE",
  UPDATE_ORDER = "UPDATE_ORDER",
  UPDATE_ORDER_SUCCESS = "UPDATE_ORDER_SUCCESS",
  UPDATE_ORDER_FAILURE = "UPDATE_ORDER_FAILURE",
  DELETE_ORDER = "DELETE_ORDER",
  DELETE_ORDER_SUCCESS = "DELETE_ORDER_SUCCESS",
  DELETE_ORDER_FAILURE = "DELETE_ORDER_FAILURE",
  CREATE_ORDER_RETURN = "CREATE_ORDER_RETURN",
  CREATE_ORDER_RETURN_SUCCESS = "CREATE_ORDER_RETURN_SUCCESS",
  CREATE_ORDER_RETURN_FAILURE = "CREATE_ORDER_RETURN_FAILURE",
  GET_ORDER_AND_RETURN_BY_ID = "GET_ORDER_AND_RETURN_BY_ID",
  GET_ORDER_AND_RETURN_BY_ID_SUCCESS = "GET_ORDER_AND_RETURN_BY_ID_SUCCESS",
  GET_ORDER_AND_RETURN_BY_ID_FAILURE = "GET_ORDER_AND_RETURN_BY_ID_FAILURE",
  GET_ORDER_BY_REFERENCE = "GET_ORDER_BY_REFERENCE",
  GET_ORDER_BY_REFERENCE_SUCCESS = "GET_ORDER_BY_REFERENCE_SUCCESS",
  GET_ORDER_BY_REFERENCE_FAILURE = "GET_ORDER_BY_REFERENCE_FAILURE",
  GET_ORDER_BY_REFERENCE_FILTER = "GET_ORDER_BY_REFERENCE_FILTER",
  GET_ORDER_BY_REFERENCE_FILTER_SUCCESS = "GET_ORDER_BY_REFERENCE_FILTER_SUCCESS",
  GET_ORDER_BY_REFERENCE_FILTER_FAILURE = "GET_ORDER_BY_REFERENCE_FILTER_FAILURE",
}